modified end points to one class
